+++ packages/angular-sdk/projects/zoom-clone/src/environments/environment.prod.ts
@@ -1,4 +1,0 @@
-export const environment = {
-  production: true,
-  apiKey: process.env['STREAM_API_KEY'] || '5mxvmc2t4qys',
-};